WIP implementing the Bongo cat animation.
The bongo cat get rendered on the right half of the OLED but the tapping
animation and WPM is not working at all. Then, another issue is that
when the defined 5 minutes idle time reached, the bongo cat OLED will be
turned off and not get back on when the keyboard being used again.